# Optimize HKDF implementation.