--- a/src/Quidjibo.DataProtection/Quidjibo.DataProtection.csproj
+++ b/src/Quidjibo.DataProtection/Quidjibo.DataProtection.csproj
@@ -13,10 +13,14 @@
     <Authors>Smiggleworth</Authors>
     <Company>Smiggleworth</Company>
     <Version>1.0.0</Version>
+    <LangVersion>7.3</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Quidjibo\Quidjibo.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="System.Memory" Version="4.5.1" />
   </ItemGroup>
 
 </Project>