@vcsjones 2018-10-26
If you're curious, here is `netcoreapp2.1`-ified.

``` ini

BenchmarkDotNet=v0.11.1, OS=macOS Mojave 10.14 (18A391) [Darwin 18.0.0]
Intel Core i7-6567U CPU 3.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.403
  [Host]     : .NET Core 2.1.5 (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.5 (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT


```
| Method | ExpandSize | Algorithm |     Mean |     Error |    StdDev |   Median |  Gen 0 | Allocated |
|------- |----------- |---------- |---------:|----------:|----------:|---------:|-------:|----------:|
| **Expand** |         **32** |      **SHA1** | **2.729 us** | **0.0575 us** | **0.1107 us** | **2.679 us** | **0.1297** |     **272 B** |
| **Expand** |         **32** |    **SHA256** | **2.236 us** | **0.0489 us** | **0.1394 us** | **2.206 us** | **0.1297** |     **280 B** |
| **Expand** |         **32** |    **SHA512** | **2.671 us** | **0.0532 us** | **0.1098 us** | **2.671 us** | **0.1450** |     **312 B** |
| **Expand** |         **64** |      **SHA1** | **4.783 us** | **0.0934 us** | **0.1535 us** | **4.794 us** | **0.1297** |     **272 B** |
| **Expand** |         **64** |    **SHA256** | **3.611 us** | **0.0740 us** | **0.2124 us** | **3.575 us** | **0.1297** |     **280 B** |
| **Expand** |         **64** |    **SHA512** | **2.750 us** | **0.0677 us** | **0.1953 us** | **2.676 us** | **0.1450** |     **312 B** |
| **Expand** |        **128** |      **SHA1** | **7.840 us** | **0.1562 us** | **0.2385 us** | **7.811 us** | **0.1297** |     **272 B** |
| **Expand** |        **128** |    **SHA256** | **6.457 us** | **0.1343 us** | **0.3539 us** | **6.300 us** | **0.1297** |     **280 B** |
| **Expand** |        **128** |    **SHA512** | **4.326 us** | **0.0831 us** | **0.0924 us** | **4.286 us** | **0.1450** |     **312 B** |