@johnallers 2018-10-26
This looks great. Thanks @vcsjones!

I'm okay with this PR, if you are @smiggleworth